Fix distance errors (GeoDistance)
Don't serialize the `DistanceMatrix` to JSON. Instead, it is more reliable to work with the Java objects (e.g. to avoid exceptions like "JSON key not found").

Additionally: Removing " km" from the string is highly unreliable! `distance.humanReadable` is not always given in kilometers (e.g. if distance is < 100 meters).
Luckily, an exception is thrown (`NumberFormatException` by `Double.parseDouble()`) in that case; otherwise, this edge case would be unnoticed during runtime.